backend/requirements.txt
Django==5.0.7
djangorestframework==3.15.2
psycopg[binary]==3.2.1
gunicorn==22.0.0
celery==5.4.0
redis==5.0.7
python-dotenv==1.0.1
django-cors-headers==4.4.0
django-cors-headers==4.4.0
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.1
